pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegra"
version = "0.1.0"
description = "Open source LangGraph Platform alternative - Self-hosted AI agent backend with FastAPI and PostgreSQL. Zero vendor lock-in, full control over your agent infrastructure."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["langgraph", "ai-agents", "self-hosted", "agent-backend", "fastapi", "postgresql", "langchain", "agent-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
authors = [
    {name = "Muhammad Ibrahim", email = "mibrahim37612@gmail.com"}
]
maintainers = [
    {name = "Muhammad Ibrahim", email = "mibrahim37612@gmail.com"}
]
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "langchain-openai>=0.3.28",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.6.0",
    "langchain>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.35.0",
    "langfuse>=3.3.4",
    "tavily-python>=0.3.0",
    "httpx>=0.24.0",
    "pydantic-settings>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/ibbybuilds/aegra"
Repository = "https://github.com/ibbybuilds/aegra"
Documentation = "https://github.com/ibbybuilds/aegra#readme"
"Bug Tracker" = "https://github.com/ibbybuilds/aegra/issues"
"Feature Requests" = "https://github.com/ibbybuilds/aegra/issues"



[tool.hatch.build.targets.wheel]
packages = ["src/agent_server"]

[tool.hatch.build.targets.wheel.sources]
"graphs/react_agent" = "react_agent"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "e2e: marks tests as end-to-end tests",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]
